--- a/2024/day11/main.go
+++ b/2024/day11/main.go
@@ -1,0 +1,63 @@
+package main
+
+import (
+	_ "embed"
+	"flag"
+	"fmt"
+	"log/slog"
+	"strings"
+
+	"github.com/Javinator9889/aoc-2024/cast"
+	"github.com/Javinator9889/aoc-2024/util"
+)
+
+//go:embed input.txt
+var input string
+
+func init() {
+	// do this in init (not main) so test file has same input
+	input = strings.TrimRight(input, "\n")
+	if len(input) == 0 {
+		panic("empty input.txt file")
+	}
+}
+
+func main() {
+	var part int
+	var debug bool
+	flag.IntVar(&part, "part", 1, "part 1 or 2")
+	flag.BoolVar(&debug, "debug", false, "debug mode")
+	flag.Parse()
+	fmt.Println("Running part", part)
+	if debug {
+		slog.SetLogLoggerLevel(slog.LevelDebug)
+	}
+
+	if part == 1 {
+		ans := part1(input)
+		util.CopyToClipboard(fmt.Sprintf("%v", ans))
+		fmt.Println("Output:", ans)
+	} else {
+		ans := part2(input)
+		util.CopyToClipboard(fmt.Sprintf("%v", ans))
+		fmt.Println("Output:", ans)
+	}
+}
+
+func part1(input string) int {
+	parsed := parseInput(input)
+	_ = parsed
+
+	return 0
+}
+
+func part2(input string) int {
+	return 0
+}
+
+func parseInput(input string) (ans []int) {
+	for _, line := range strings.Split(input, "\n") {
+		ans = append(ans, cast.ToInt(line))
+	}
+	return ans
+}