--- a/2024/day11/main.go
+++ b/2024/day11/main.go
@@ -44,20 +44,77 @@ func main() {
 	}
 }
 
-func part1(input string) int {
-	parsed := parseInput(input)
-	_ = parsed
+type Stone struct {
+	value int
+	next  *Stone
+}
 
-	return 0
+func (s Stone) Len() int {
+	return len(cast.ToString(s.value))
+}
+
+func (s *Stone) Blink() (next *Stone) {
+	switch {
+	case s.value == 0:
+		s.value = 1
+		next = s.next
+	case s.Len()%2 == 0:
+		sv := cast.ToString(s.value)
+		first, second := sv[:len(sv)/2], sv[len(sv)/2:]
+		s.value = cast.ToInt(first)
+		newStone := &Stone{value: cast.ToInt(second)}
+		s.next, newStone.next = newStone, s.next
+		next = newStone.next // Skip the new stone
+	default:
+		s.value *= 2024
+		next = s.next
+	}
+	return
+}
+
+func (s *Stone) Size() (size int) {
+	for stone := s; stone != nil; stone = stone.next {
+		size++
+	}
+	return
+}
+
+func (s *Stone) String() string {
+	var sb strings.Builder
+	for stone := s; stone != nil; stone = stone.next {
+		sb.WriteString(fmt.Sprintf("%d ", stone.value))
+	}
+	return sb.String()
+}
+
+func part1(input string) int {
+	stones := parseInput(input)
+	slog.Debug("Stones:", "stones", stones)
+	for i := 0; i < 25; i++ {
+		for stone := stones; stone != nil; stone = stone.Blink() {
+		}
+	}
+
+	return stones.Size()
 }
 
 func part2(input string) int {
 	return 0
 }
 
-func parseInput(input string) (ans []int) {
+func parseInput(input string) (stones *Stone) {
+	var prev *Stone
 	for _, line := range strings.Split(input, "\n") {
-		ans = append(ans, cast.ToInt(line))
+		for _, num := range strings.Fields(line) {
+			if prev == nil {
+				prev = &Stone{value: cast.ToInt(num)}
+				stones = prev
+				continue
+			}
+			newStone := &Stone{value: cast.ToInt(num)}
+			prev.next = newStone
+			prev = newStone
+		}
 	}
-	return ans
+	return
 }