--- a/2024/day11/main_test.go
+++ b/2024/day11/main_test.go
@@ -40,11 +40,11 @@ func Test_part2(t *testing.T) {
 		input string
 		want  int
 	}{
-		// {
-		// 	name:  "actual",
-		// 	input: input,
-		// 	want:  0,
-		// },
+		{
+			name:  "actual",
+			input: input,
+			want:  220566831337810,
+		},
 	}
 	slog.SetLogLoggerLevel(slog.LevelDebug)
 	for _, tt := range tests {
@@ -53,5 +53,30 @@ func Test_part2(t *testing.T) {
 				t.Errorf("part2() = %v, want %v", got, tt.want)
 			}
 		})
+	}
+}
+
+func recursiveImpl(stones *Stone, n int) int {
+	return blink(n, stones)
+}
+
+func cachedImpl(stones *Stone, n int) (count int) {
+	for st := stones; st != nil; st = st.next {
+		count += st.Stones(n)
+	}
+	return
+}
+
+func BenchmarkRecursive(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		stones := parseInput(input)
+		recursiveImpl(stones, 25)
+	}
+}
+
+func BenchmarkCached(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		stones := parseInput(input)
+		cachedImpl(stones, 25)
 	}
 }