--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/Javinator9889/aoc-2024
 
-go 1.23
+go 1.23.3
 
 require (
 	golang.org/x/net v0.31.0 // indirect