… 2024/day11/input.txt