I had some troubles initially as the iterative implementation, although iterative, was causing lots of troubles with memory - I read in the Reddit subforum that you'd need 8000TB to store all the rocks 🤯

I optimized the implementation, going with dynamic programming. The problem now lies in the fact that running it takes almost forever 😕. And I said to myself: "For sure there are repeated values, why do it twice?" So I implemented a cache.

With the cache, it's blazing fast ⚡. It takes less than `330ms` to compute everything. So yes, for sure there were repeated values 🤣. I have created a benchmark suite for the recursive implementation and the cached one, yielding the following results:

```shell
❯ go test -benchmem -run=^$ -bench ^BenchmarkRecursive$ github.com/Javinator9889/aoc-2024/2024/day11
goos: linux
goarch: amd64
pkg: github.com/Javinator9889/aoc-2024/2024/day11
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkRecursive-8          21          52940547 ns/op        12642919 B/op    1021070 allocs/op
PASS
ok      github.com/Javinator9889/aoc-2024/2024/day11    1.169s
```

Whereas the cached version yields:

```shell
❯ go test -benchmem -run=^$ -bench ^BenchmarkCached$ github.com/Javinator9889/aoc-2024/2024/day11
goos: linux
goarch: amd64
pkg: github.com/Javinator9889/aoc-2024/2024/day11
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkCached-8            339           3280275 ns/op          972302 B/op      46783 allocs/op
PASS
ok      github.com/Javinator9889/aoc-2024/2024/day11    1.471s
```

Using `benchstat` tool, the difference is quite clear:

```shell
❯ benchstat recursive cached
goos: linux
goarch: amd64
pkg: github.com/Javinator9889/aoc-2024/2024/day11
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
            │  recursive  │       cached        │
            │   sec/op    │    sec/op     vs base   │
Recursive-8   52.08m ± 3%
Cached-8                    3.423m ± 14%
geomean       52.08m        3.423m        ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

            │  recursive   │       cached        │
            │     B/op     │     B/op      vs base   │
Recursive-8   12.06Mi ± 0%
Cached-8                     949.5Ki ± 0%
geomean       12.06Mi        949.5Ki       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean

            │  recursive  │       cached       │
            │  allocs/op  │  allocs/op   vs base   │
Recursive-8   1.021M ± 0%
Cached-8                    46.78k ± 0%
geomean       1.021M        46.78k       ? ¹ ²
¹ benchmark set differs from baseline; geomeans may not be comparable
² ratios must be >0 to compute geomean
```